Abstract out a query function
